+++ pymodbus/payload.py
@@ -10,6 +10,7 @@ from pymodbus.interfaces import IPayloadBuilder
 from pymodbus.constants import Endian
 from pymodbus.utilities import pack_bitstring
 from pymodbus.utilities import unpack_bitstring
+from pymodbus.utilities import make_byte_string
 from pymodbus.exceptions import ParameterException
 
 
@@ -30,7 +31,7 @@ class BinaryPayloadBuilder(IPayloadBuilder):
     def __init__(self, payload=None, endian=Endian.Little):
         ''' Initialize a new instance of the payload builder
 
-        :param payload: Raw payload data to initialize with
+        :param payload: Raw binary payload data to initialize with
         :param endian: The endianess of the payload
         '''
         self._payload = payload or []
@@ -175,6 +176,7 @@ class BinaryPayloadBuilder(IPayloadBuilder):
 
         :param value: The value to add to the buffer
         '''
+        value = make_byte_string(value)
         fstring = self._endian + str(len(value)) + 's'
         self._payload.append(pack(fstring, value))
 
@@ -246,14 +248,16 @@ class BinaryPayloadDecoder(object):
         self._pointer += 1
         fstring = self._endian + 'B'
         handle = self._payload[self._pointer - 1:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_bits(self):
         ''' Decodes a byte worth of bits from the buffer
         '''
         self._pointer += 1
-        fstring = self._endian + 'B'
+        # fstring = self._endian + 'B'
         handle = self._payload[self._pointer - 1:self._pointer]
+        handle = make_byte_string(handle)
         return unpack_bitstring(handle)
 
     def decode_16bit_uint(self):
@@ -262,6 +266,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 2
         fstring = self._endian + 'H'
         handle = self._payload[self._pointer - 2:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_32bit_uint(self):
@@ -270,6 +275,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 4
         fstring = self._endian + 'I'
         handle = self._payload[self._pointer - 4:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_64bit_uint(self):
@@ -278,6 +284,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 8
         fstring = self._endian + 'Q'
         handle = self._payload[self._pointer - 8:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_8bit_int(self):
@@ -286,6 +293,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 1
         fstring = self._endian + 'b'
         handle = self._payload[self._pointer - 1:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_16bit_int(self):
@@ -294,6 +302,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 2
         fstring = self._endian + 'h'
         handle = self._payload[self._pointer - 2:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_32bit_int(self):
@@ -302,6 +311,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 4
         fstring = self._endian + 'i'
         handle = self._payload[self._pointer - 4:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_64bit_int(self):
@@ -310,6 +320,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 8
         fstring = self._endian + 'q'
         handle = self._payload[self._pointer - 8:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_32bit_float(self):
@@ -318,6 +329,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 4
         fstring = self._endian + 'f'
         handle = self._payload[self._pointer - 4:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_64bit_float(self):
@@ -326,6 +338,7 @@ class BinaryPayloadDecoder(object):
         self._pointer += 8
         fstring = self._endian + 'd'
         handle = self._payload[self._pointer - 8:self._pointer]
+        handle = make_byte_string(handle)
         return unpack(fstring, handle)[0]
 
     def decode_string(self, size=1):